+++ apps/exp.py
@@ -1,244 +1,425 @@
-"""
-SALAMI Explorer - Streamlit App using BNL
-
-An interactive web application for exploring the SALAMI music dataset with audio playback,
-visualization, and metadata browsing; powered by BNL.
-
-Features:
-- Browse 1,400+ music tracks with rich metadata (title, artist, duration)
-- Stream MP3 audio directly from cloud storage (Cloudflare R2)
-- Real-time waveform and MFCC visualization
-- Dual data source support: Cloud (default) or local filesystem
-- Cached data loading for optimal performance
-
-Usage:
-    pixi run exp
-
-Data Sources:
-- Cloud: Streams data directly from R2 bucket using online manifest
-- Local: Supports SALAMI-style datasets with a path-based manifest containing relative asset paths
-
-Architecture:
-- Cloud-native: Streams data from R2 bucket without local storage
-- Efficient caching: Uses Streamlit's @st.cache_data and @st.cache_resource
-- Robust error handling: Graceful fallbacks for missing assets
-- Real-time analysis: On-demand audio processing with librosa
-"""
-
 import os
-from typing import Any
 
-import librosa
-import librosa.display
-import matplotlib.pyplot as plt
+import numpy as np
+import plotly.graph_objects as go
 import streamlit as st
+from plotly.subplots import make_subplots
+from scipy.signal import find_peaks
+from sklearn.neighbors import KernelDensity
 
 import bnl
+from bnl import ops
 
-# --- Page and Data Configuration ---
+
+# --- Core Computational Class (Reused for both stages) ---
+class KDEBoundaryExplorer:
+    """
+    Manages the state and computation for KDE-based boundary merging.
+    This version is headless and optimized for use in a web app.
+    """
+
+    def __init__(self, data_points: np.ndarray, weights: np.ndarray = None, viz_resolution: int = 2000):
+        if data_points.size == 0:
+            # Return a non-functional but safe object if there's no data
+            self.resampled_times = np.array([])
+            self.viz_grid = np.array([])
+            self.density = np.array([])
+            self.peaks = np.array([])
+            self.peak_saliences = np.array([])
+            return
+
+        # --- Resample ONCE based on weights ---
+        if weights is not None and weights.sum() > 0:
+            probabilities = weights / np.sum(weights)
+            resample_size = min(len(data_points) * 10, 5000)
+            self.resampled_times = np.random.choice(data_points, size=resample_size, p=probabilities).reshape(-1, 1)
+        else:  # Handle unweighted case
+            self.resampled_times = data_points.reshape(-1, 1)
+
+        # --- Setup for visualization grid ---
+        min_t, max_t = data_points.min(), data_points.max()
+        # Add a small buffer to avoid plotting issues at the edges
+        buffer = (max_t - min_t) * 0.05 if (max_t - min_t) > 0 else 0.1
+        self.viz_grid = np.linspace(min_t - buffer, max_t + buffer, viz_resolution).reshape(-1, 1)
+
+        self.density = np.zeros(self.viz_grid.shape)
+        self.peaks = np.array([])
+        self.peak_saliences = np.array([])
+
+    def update_bandwidth(self, bandwidth: float):
+        """
+        Updates the KDE and finds new peaks for the given bandwidth.
+        """
+        if bandwidth <= 0 or self.resampled_times.size == 0:
+            return
+
+        kde = KernelDensity(kernel="gaussian", bandwidth=bandwidth).fit(self.resampled_times)
+        log_density = kde.score_samples(self.viz_grid)
+        self.density = np.exp(log_density)
+
+        peak_indices, _ = find_peaks(
+            self.density, prominence=np.max(self.density) * 0.01 if self.density.size > 0 else 0.01
+        )
+
+        self.peaks = self.viz_grid.flatten()[peak_indices]
+        self.peak_saliences = self.density[peak_indices]
+
+    def get_plot_data(self) -> dict:
+        """Returns all necessary data for plotting."""
+        return {
+            "grid": self.viz_grid.flatten(),
+            "density": self.density,
+            "peaks": self.peaks,
+            "peak_saliences": self.peak_saliences,
+        }
+
+
+# --- Streamlit Application ---
+
 st.set_page_config(layout="wide")
-st.title("SALAMI Explorer")
-
-# --- Data Source and Path Configuration ---
-R2_BUCKET_PUBLIC_URL = "https://pub-05e404c031184ec4bbf69b0c2321b98e.r2.dev"
-CLOUD_MANIFEST_URL = f"{R2_BUCKET_PUBLIC_URL}/manifest_cloud_boolean.csv"
-LOCAL_MANIFEST_PATH = os.path.expanduser("~/data/salami/metadata.csv")
-
-# Let user choose data source
-data_source_option = st.sidebar.radio(
-    "Select Data Source:",
-    ("Cloud (R2)", "Local Filesystem"),
-)
-
-# Manual session state management for data source
-if "data_source_choice" not in st.session_state:
-    st.session_state.data_source_choice = data_source_option
-elif data_source_option != st.session_state.data_source_choice:
-    st.session_state.data_source_choice = data_source_option
-    # Clear everything when data source changes
-    preserve_keys = ["data_source_choice"]  # Add other critical keys here if needed
-    for key in list(st.session_state.keys()):
-        if key not in preserve_keys:
-            del st.session_state[key]
-    st.cache_data.clear()
-    st.rerun()
+st.sidebar.title("Hierarchical KDE Boundary Explorer")
 
 
-# --- Data Loading Functions (Cached for performance) ---
+# --- Data Loading ---
 @st.cache_resource
-def get_dataset(source_type: str) -> bnl.data.Dataset:
-    """Loads and caches the dataset object based on selected source."""
-    if source_type == "Cloud (R2)":
-        try:
-            # Load from the online cloud manifest - Dataset will auto-detect it's a cloud source
-            return bnl.data.Dataset(CLOUD_MANIFEST_URL)
-        except Exception as e:
-            st.error(f"Failed to load cloud manifest from: {CLOUD_MANIFEST_URL}")
-            st.error(f"Error: {e}")
-            st.error("Please check your internet connection and verify the cloud service is available.")
-            st.stop()
-
-    elif source_type == "Local Filesystem":
-        try:
-            # The Dataset class now infers the source type from the local path
-            return bnl.data.Dataset(LOCAL_MANIFEST_PATH)
-        except Exception as e:
-            st.error(f"Failed to load LOCAL manifest from: {LOCAL_MANIFEST_PATH} \n Error: {e}")
-            st.stop()
-    else:
-        st.error("Invalid data source selected.")
-        st.stop()
-
-
-# Get the current dataset based on user's choice
-current_dataset = get_dataset(st.session_state.data_source_choice)
-
-
-@st.cache_data
-def load_track_data(_dataset: bnl.data.Dataset, track_id: str) -> tuple:
-    """Loads track data and populates track.info with metadata like title/artist."""
-    track = _dataset[track_id]
-
-    # Load audio using the Track's built-in method
-    waveform, sr = track.load_audio()
-
-    return track, waveform, sr
-
-
-@st.cache_data
-def create_audio_analysis_plot(waveform: Any, sr: Any) -> plt.Figure | None:
-    """Creates a matplotlib figure with waveform and MFCC plots."""
-    if waveform is None or sr is None:
+def get_dataset(manifest_path="~/data/salami/metadata.csv"):
+    """Loads the dataset object, returns None if not found."""
+    manifest_path = os.path.expanduser(manifest_path)
+    if not os.path.exists(manifest_path):
+        st.sidebar.error(f"Manifest not found: {manifest_path}")
+        return None
+    try:
+        return bnl.data.Dataset(manifest_path=manifest_path)
+    except Exception as e:
+        st.sidebar.error(f"Error loading dataset: {e}")
         return None
 
-    fig, (ax1, ax2) = plt.subplots(2, 1, figsize=(6, 4), sharex=True, constrained_layout=True)
-    librosa.display.waveshow(waveform, sr=sr, ax=ax1)
-    ax1.set_title("Waveform")
-    ax1.set_ylabel("Amplitude")
-    ax1.set_xlabel(None)
 
-    mfccs = librosa.feature.mfcc(y=waveform, sr=sr, n_mfcc=40)
-    img = librosa.display.specshow(mfccs, x_axis="time", y_axis="mel", sr=sr, ax=ax2)
-    ax2.set_title("MFCC")
+@st.cache_resource
+def get_dataset_cloud():
+    r2_url = "https://pub-05e404c031184ec4bbf69b0c2321b98e.r2.dev"
+    try:
+        return bnl.data.Dataset(manifest_path=f"{r2_url}/manifest_cloud_boolean.csv")
+    except Exception as e:
+        st.sidebar.error(f"Error loading dataset: {e}")
+        return None
 
-    fig.colorbar(img, ax=ax2, format="%+2.0f dB")
+
+@st.cache_data
+def load_boundary_contour(_dataset, track_id):
+    """Loads a specific track's boundary contour."""
+    if _dataset is None or track_id is None:
+        return None
+    try:
+        track = _dataset[track_id]
+        est = track.load_annotation("adobe-mu1gamma9")
+        return ops.boundary_salience(est, strategy="depth")
+    except Exception as e:
+        st.warning(f"Failed to load data for track {track_id}: {e}")
+        return None
+
+
+slm_ds = get_dataset_cloud()
+
+# --- Sidebar Controls ---
+track_id = st.sidebar.selectbox("Select Track ID", slm_ds.track_ids, index=8)
+boundary_contour = load_boundary_contour(slm_ds, track_id)
+initial_times = np.array([b.time for b in boundary_contour.boundaries])
+initial_saliences = np.array([b.salience for b in boundary_contour.boundaries])
+
+st.sidebar.write("### Stage 1: Time Grouping")
+# Configuration for Time Slider (Log Scale)
+time_bw_min, time_bw_max, time_bw_default = 0.1, 10.0, 1.0
+time_bw_min_log, time_bw_max_log = np.log10(time_bw_min), np.log10(time_bw_max)
+time_bw_options = np.logspace(time_bw_min_log, time_bw_max_log, num=50)
+# Find the closest default value in the new options to ensure it's a valid choice
+time_bw_val = min(time_bw_options, key=lambda x: abs(x - time_bw_default))
+
+time_bandwidth = st.sidebar.select_slider(
+    "Time KDE Bandwidth (σ, Log Scale)",
+    options=time_bw_options,
+    value=time_bw_val,
+    format_func=lambda x: f"{x:.3f} s",
+    help="Controls the smoothness of the temporal density estimate. The slider is on a log scale for finer control.",
+)
+
+
+st.sidebar.write("### Stage 2: Salience Quantization")
+# Configuration for Salience Slider (Log Scale)
+sal_bw_min, sal_bw_max, sal_bw_default = 1e-4, 1e-2, 1e-3
+sal_bw_min_log, sal_bw_max_log = np.log10(sal_bw_min), np.log10(sal_bw_max)
+sal_bw_options = np.logspace(sal_bw_min_log, sal_bw_max_log, num=50)
+# Find the closest default value in the new options to ensure it's a valid choice
+sal_bw_val = min(sal_bw_options, key=lambda x: abs(x - sal_bw_default))
+
+salience_bandwidth = st.sidebar.select_slider(
+    "Salience KDE Bandwidth (Log Scale)",
+    options=sal_bw_options,
+    value=sal_bw_val,
+    format_func=lambda x: f"{x:.5f}",  # Format for readability
+    help="Controls the smoothness of the salience density estimate. The slider is on a log scale for finer control.",
+)
+
+
+# --- App Description ---
+st.sidebar.markdown("---")
+st.sidebar.info(
+    """
+This app performs a two-stage analysis.
+The main plot merges temporal boundaries using a KDE.
+The narrow plot on the right acts as a marginal histogram,
+taking the *saliences* of the merged peaks and performing a second KDE to find significant salience levels.
+"""
+)
+
+
+# --- Core Computations ---
+@st.cache_resource
+def get_time_explorer(times, saliences):
+    return KDEBoundaryExplorer(times, saliences)
+
+
+time_explorer = get_time_explorer(initial_times, initial_saliences)
+
+# --- Core Logic ---
+# Stage 1: Analyze time points
+time_explorer.update_bandwidth(time_bandwidth)
+time_plot_data = time_explorer.get_plot_data()
+resulting_saliences = time_plot_data["peak_saliences"]
+
+# Stage 2: Analyze the saliences from stage 1
+salience_plot_data = {
+    "grid": np.array([]),
+    "density": np.array([]),
+    "peaks": np.array([]),
+    "peak_saliences": np.array([]),
+}
+quantized_saliences = np.array([])
+
+if resulting_saliences.size > 0:
+    salience_explorer = KDEBoundaryExplorer(resulting_saliences)
+    salience_explorer.update_bandwidth(salience_bandwidth)
+    salience_plot_data = salience_explorer.get_plot_data()
+    salience_levels = salience_plot_data["peaks"]
+
+    if salience_levels.size > 0:
+        closest_level_indices = np.argmin(np.abs(salience_levels[:, np.newaxis] - resulting_saliences), axis=0)
+        quantized_saliences = salience_levels[closest_level_indices]
+
+
+# --- Plotting Functions ---
+def make_interactive_plot():
+    """Creates the main interactive plot with two subplots."""
+    fig = make_subplots(rows=1, cols=2, shared_yaxes=True, column_widths=[0.8, 0.2], horizontal_spacing=0.02)
+
+    # --- Traces for Main Plot (Time) ---
+    fig.add_trace(go.Scatter(name="Time KDE Density", line=dict(color="royalblue")), row=1, col=1)
+    fig.add_trace(
+        go.Scatter(name="Original Peaks", mode="markers", marker=dict(color="grey", size=8, symbol="diamond-open")),
+        row=1,
+        col=1,
+    )
+    fig.add_trace(
+        go.Scatter(name="Quantized Peaks", mode="markers", marker=dict(color="red", size=6, symbol="circle")),
+        row=1,
+        col=1,
+    )
+    fig.add_trace(
+        go.Scatter(
+            name="Time Data Rug",
+            mode="markers",
+            marker_symbol="line-ns",
+            marker=dict(color="rgba(170, 170, 170, 0.8)", size=12, line_width=1.5),
+            showlegend=False,
+        ),
+        row=1,
+        col=1,
+    )
+
+    # --- Traces for Marginal Plot (Salience) ---
+    fig.add_trace(go.Scatter(name="Salience KDE Density", line=dict(color="mediumseagreen")), row=1, col=2)
+    fig.add_trace(
+        go.Scatter(name="Salience Levels", mode="markers", marker=dict(color="red", size=6, symbol="circle")),
+        row=1,
+        col=2,
+    )
+    fig.add_trace(
+        go.Scatter(
+            name="Salience Data Rug",
+            mode="markers",
+            marker_symbol="line-ew",
+            marker=dict(color="rgba(170, 170, 170, 0.8)", size=12, line_width=1.5),
+            showlegend=False,
+        ),
+        row=1,
+        col=2,
+    )
+
+    # --- Layout ---
+    fig.update_layout(
+        title_text="Stage 1: Temporal Merging (Left) & Stage 2: Salience Quantization (Right)",
+        height=450,  # Reduced height
+        margin=dict(l=20, r=20, t=60, b=20),
+        legend=dict(yanchor="top", y=0.98, xanchor="left", x=0.01),
+        xaxis=dict(zeroline=False),
+        yaxis=dict(zeroline=False),
+        xaxis2=dict(zeroline=False),
+    )
+    fig.update_xaxes(title_text="Time", row=1, col=1)
+    fig.update_yaxes(title_text="Density / Salience", row=1, col=1)
+    fig.update_xaxes(title_text="Density", row=1, col=2)
     return fig
 
 
-# --- UI Layout & App State ---
-def reset_track_state() -> None:
-    """Resets flags when a new track is selected, triggering a data reload."""
-    st.session_state.track_loaded = False
-    st.session_state.is_stabilized = False
+@st.cache_data
+def compute_persistence_data(p_initial_times, p_initial_saliences, time_bw_range, sal_bw_range):
+    """
+    Pre-computes the number of peaks across the entire parameter space.
+    This function is cached; it will re-run only when the track's data changes.
+    """
+    # Use a local explorer instance for this computation, based on the specific track's data
+    _time_explorer = KDEBoundaryExplorer(p_initial_times, p_initial_saliences)
+
+    time_persistence = []
+    salience_persistence = np.zeros((len(sal_bw_range), len(time_bw_range)))
+
+    for i, t_bw in enumerate(time_bw_range):
+        _time_explorer.update_bandwidth(t_bw)
+        time_peaks = _time_explorer.get_plot_data()["peaks"]
+        time_persistence.append(len(time_peaks))
+
+        saliences = _time_explorer.get_plot_data()["peak_saliences"]
+        if saliences.size > 0:
+            # OPTIMIZATION: Initialize explorer once per time-bandwidth setting
+            salience_explorer = KDEBoundaryExplorer(saliences)
+            for j, s_bw in enumerate(sal_bw_range):
+                salience_explorer.update_bandwidth(s_bw)
+                salience_peaks = salience_explorer.get_plot_data()["peaks"]
+                salience_persistence[j, i] = len(salience_peaks)
+
+    return time_bw_range, sal_bw_range, time_persistence, salience_persistence
 
 
-# Sidebar for parameter tuning.
-# Define this before the logic that uses its state.
-with st.sidebar:
-    # This selectbox initializes `st.session_state.track_id` on the first run.
-    if not current_dataset.track_ids:
-        st.warning("No track IDs found in the selected dataset.")
-        st.stop()
-
-    st.selectbox(
-        "Select a SALAMI track id:",
-        current_dataset.track_ids,
-        key="track_id",
-        on_change=reset_track_state,
+def make_persistence_plot(
+    time_bw_range, sal_bw_range, time_persistence, salience_persistence, current_time_bw, current_sal_bw
+):
+    """Creates the persistence landscape and heatmap, arranged vertically to share the x-axis."""
+    fig = make_subplots(
+        rows=2,
+        cols=1,
+        shared_xaxes=True,
+        vertical_spacing=0.1,
+        row_heights=[0.3, 0.7],
+        subplot_titles=(
+            "Number of Temporal Boundaries",
+            "Number of Salience Levels",
+        ),
     )
-    # Initialize track_id to first available if not set or invalid
-    if st.session_state.get("track_id") not in current_dataset.track_ids:
-        st.session_state.track_id = current_dataset.track_ids[0]
-        reset_track_state()
 
-    # Show current status
-    st.write(f"📊 **Data source:** {st.session_state.data_source_choice}")
-    st.write(f"🎯 **Track:** {st.session_state.track_id}")
+    # Plot 1 (Top): Time Persistence (Number of Boundaries vs. Time Bandwidth)
+    fig.add_trace(
+        go.Scatter(x=time_bw_range, y=time_persistence, mode="lines", name="Time Peaks", showlegend=False),
+        row=1,
+        col=1,
+    )
+    fig.add_vline(x=current_time_bw, line_width=2, line_dash="dash", line_color="red", name="Time BW", row=1, col=1)
+    fig.update_yaxes(title_text="Num Boundaries", row=1, col=1)
 
-    if st.session_state.get("track_loaded", False) and hasattr(st.session_state, "track"):
-        tinfo = st.session_state.track.info
-        st.write(f"Title: {tinfo.get('title', 'N/A')}")  # Use .get for safety
-        st.write(f"Artist: {tinfo.get('artist', 'N/A')}")
+    # Plot 2 (Bottom): Salience Persistence Heatmap
+    fig.add_trace(
+        go.Heatmap(
+            z=salience_persistence,
+            x=time_bw_range,
+            y=sal_bw_range,
+            colorscale="Viridis",
+            showscale=True,
+            colorbar=dict(title="Num Levels"),
+        ),
+        row=2,
+        col=1,
+    )
+    # Add cursors to the heatmap
+    fig.add_vline(x=current_time_bw, line_width=2, line_dash="dash", line_color="red", name="Time BW", row=2, col=1)
+    fig.add_hline(y=current_sal_bw, line_width=2, line_dash="dash", line_color="red", name="Salience BW", row=2, col=1)
 
-    st.header("Tuning Parameters (Placeholder)")
-    bw_peak = st.slider("Bandwidth for Peak Picking", 0.1, 5.0, 1.0, 0.1)
-    bw_group = st.slider("Bandwidth for Depth Grouping", 0.1, 5.0, 1.0, 0.1)
-
-
-# --- Core Logic ---
-# Initialize state for first run if they don't exist.
-if "track_loaded" not in st.session_state:
-    st.session_state.track_loaded = False
-if "is_stabilized" not in st.session_state:  # May not be needed anymore
-    st.session_state.is_stabilized = False
-
-
-# Load track data when needed
-if not st.session_state.track_loaded:
-    with st.spinner(f"Loading track {st.session_state.track_id}..."):
-        track, waveform, sr = load_track_data(current_dataset, st.session_state.track_id)
-        analysis_plot = create_audio_analysis_plot(waveform, sr) if waveform is not None else None
-
-    st.session_state.track = track
-    st.session_state.analysis_plot = analysis_plot
-    st.session_state.track_loaded = True
-
-if st.session_state.track_loaded and not st.session_state.is_stabilized:
-    # Data has just been loaded. Force one immediate rerun.
-    # This "stabilizes" the UI, preventing widgets like st.audio from
-    # resetting when a slider is moved for the first time.
-    st.session_state.is_stabilized = True
-    st.rerun()
+    # Update axes and layout
+    fig.update_xaxes(title_text="Time Bandwidth (σ)", row=2, col=1, type="log")
+    fig.update_yaxes(title_text="Salience Bandwidth", row=2, col=1, type="log")
+    fig.update_xaxes(row=1, col=1, type="log")
+    fig.update_yaxes(row=1, col=1, type="log")
+    fig.update_layout(
+        title_text="Peak (left) and Level (right) vs. Bandwidth Persistence Landscape",
+        height=450,
+        showlegend=True,
+        legend=dict(orientation="h", yanchor="bottom", y=1.02, xanchor="right", x=1),
+    )
+    return fig
 
 
-# Main content area.
-if st.session_state.get("track_loaded") and hasattr(st.session_state, "track"):
-    track = st.session_state.track
-    st.header(f"Track: {track.info.get('title', st.session_state.track_id)}")
+# --- Main Display ---
 
-    # --- Audio player: find the audio path from track.info ---
-    audio_url_or_path = None
-    for key, value in track.info.items():
-        if key.startswith("audio_") and key.endswith("_path"):
-            audio_url_or_path = value
-            break
+# --- Interactive Analysis Plot ---
+if "fig" not in st.session_state:
+    st.session_state.fig = make_interactive_plot()
 
-    if audio_url_or_path:
-        st.audio(str(audio_url_or_path))  # Ensure it's a string for st.audio
+with st.session_state.fig.batch_update():
+    # --- Main plot data (Time) ---
+    st.session_state.fig.data[0].x, st.session_state.fig.data[0].y = time_plot_data["grid"], time_plot_data["density"]
+    st.session_state.fig.data[1].x, st.session_state.fig.data[1].y = (
+        time_plot_data["peaks"],
+        time_plot_data["peak_saliences"],
+    )
+    st.session_state.fig.data[2].x, st.session_state.fig.data[2].y = time_plot_data["peaks"], quantized_saliences
+    st.session_state.fig.data[3].x, st.session_state.fig.data[3].y = initial_times, np.zeros_like(initial_times)
 
-        if st.session_state.analysis_plot:
-            st.pyplot(st.session_state.analysis_plot)
-        else:
-            st.warning("Could not generate audio analysis plot (no waveform data).")
-    else:
-        st.warning(f"No audio asset found for track {st.session_state.track_id} in the manifest.")
-        st.json(track.manifest_row.to_dict())  # Display what assets the manifest says it has
+    # --- Marginal plot data (Salience) ---
+    st.session_state.fig.data[4].x, st.session_state.fig.data[4].y = (
+        salience_plot_data["density"],
+        salience_plot_data["grid"],
+    )
+    st.session_state.fig.data[5].x, st.session_state.fig.data[5].y = (
+        salience_plot_data["peak_saliences"],
+        salience_plot_data["peaks"],
+    )
+    st.session_state.fig.data[6].x, st.session_state.fig.data[6].y = (
+        np.zeros_like(resulting_saliences),
+        resulting_saliences,
+    )
 
-    # --- Annotation visualization ---
-    if track.has_annotations:
-        st.header("Annotations")
-        st.write(f"Available annotation keys: {list(track.annotations.keys())}")
-        # Attempt to load and plot the 'reference' annotation by default if it exists
-        if "reference" in track.annotations:
-            try:
-                # Using the new load_annotation method
-                ref_annotation = track.load_annotation("reference")
-                st.write(f"Loaded 'reference' annotation (type: {type(ref_annotation)}).")
-                fig, _ = ref_annotation.plot()
-                st.pyplot(fig)
-                adobe_annotation = track.load_annotation("adobe-mu1gamma9")
-                st.write(f"Loaded 'adobe-mu1gamma9' annotation (type: {type(adobe_annotation)}).")
-                adobe_fig, _ = adobe_annotation.plot()
-                st.pyplot(adobe_fig)
 
-            except Exception as e:
-                st.error(f"Error loading or plotting annotations: {e}")
-        else:
-            st.info("No 'reference' annotation key found for this track.")
-            st.write("Consider adding UI to select other annotation types and IDs if needed.")
+st.plotly_chart(st.session_state.fig, use_container_width=True)
 
-    else:
-        st.warning("No annotations found for this track.")
 
+# --- Persistence Landscape Plot ---
+# Efficiently create and update the persistence plot.
+# The expensive data computation is cached. The plot itself is stored in session
+# state, and only the crosshair positions are updated on slider interaction.
+
+# Compute or retrieve from cache
+with st.spinner("Computing persistence landscape... This may take a moment."):
+    time_bw_range, sal_bw_range, time_persistence, salience_persistence = compute_persistence_data(
+        initial_times, initial_saliences, time_bw_options, sal_bw_options
+    )
+
+# Create the plot if it's the first run or if the track has changed
+if "fig_persist" not in st.session_state or st.session_state.get("persistence_track_id") != track_id:
+    fig_persist = make_persistence_plot(
+        time_bw_range, sal_bw_range, time_persistence, salience_persistence, time_bandwidth, salience_bandwidth
+    )
+    st.session_state.fig_persist = fig_persist
+    st.session_state.persistence_track_id = track_id
 else:
-    st.warning("Please select a track or wait for data to load.")
+    fig_persist = st.session_state.fig_persist
+
+# On every interaction, retrieve the figure and just update the crosshairs
+with fig_persist.batch_update():
+    # These shapes correspond to the vline and hline calls in make_persistence_plot,
+    # assuming the order is [vline_top, vline_bottom, hline_bottom].
+    fig_persist.layout.shapes[0].x0 = time_bandwidth
+    fig_persist.layout.shapes[0].x1 = time_bandwidth
+    fig_persist.layout.shapes[1].x0 = time_bandwidth
+    fig_persist.layout.shapes[1].x1 = time_bandwidth
+    fig_persist.layout.shapes[2].y0 = salience_bandwidth
+    fig_persist.layout.shapes[2].y1 = salience_bandwidth
+
+st.plotly_chart(fig_persist, use_container_width=True)
